move around & integrate indentifier hardneing
