+++ packages/dev-env/package.json
@@ -22,6 +22,7 @@
     "@atproto/uri": "*",
     "@did-plc/lib": "^0.0.1",
     "@did-plc/server": "^0.0.1",
+    "better-sqlite3": "^7.6.2",
     "chalk": "^5.0.1",
     "dotenv": "^16.0.1",
     "get-port": "^6.1.2",
