+++ packages/identifier/README.md
@@ -1,11 +1,11 @@
-# Handle
+# Identifier
 
-Validation logic for AT handles
+Validation logic for AT identifiers - DIDs & Handles
 
 ## Usage
 
 ```typescript
-import * as handle from '@atproto/handle'
+import * as identifier from '@atproto/identifier'
 
 isValid('alice.test', ['.test']) // returns true
 ensureValid('alice.test', ['.test']) // returns void
