+++ packages/identifier/jest.config.js
@@ -2,5 +2,5 @@ const base = require('../../jest.config.base.js')
 
 module.exports = {
   ...base,
-  displayName: 'Handle',
+  displayName: 'Identifier',
 }
