+++ packages/identifier/src/index.ts
@@ -1,0 +1,4 @@
+export * from './resolve'
+export * from './handle'
+export * from './did'
+export * from './reserved'
