+++ packages/pds/tsconfig.json
@@ -12,7 +12,7 @@
     { "path": "../common/tsconfig.build.json" },
     { "path": "../crypto/tsconfig.build.json" },
     { "path": "../did-resolver/tsconfig.build.json" },
-    { "path": "../handle/tsconfig.build.json" },
+    { "path": "../identifier/tsconfig.build.json" },
     { "path": "../lexicon/tsconfig.build.json" },
     { "path": "../lex-cli/tsconfig.build.json" },
     { "path": "../repo/tsconfig.build.json" },
