+++ packages/uri/tests/uri.test.ts
@@ -1,4 +1,4 @@
-import { AtUri } from '../src/index'
+import { AtUri, ensureValidAtUri, ensureValidAtUriRegex } from '../src/index'
 
 describe('At Uris', () => {
   it('parses valid at uris', () => {
@@ -349,5 +349,151 @@ describe('At Uris', () => {
         expect(urip.hash).toBe(hash)
       }
     }
+  })
+})
+
+describe('AtUri validation', () => {
+  const expectValid = (h: string) => {
+    ensureValidAtUri(h)
+    ensureValidAtUriRegex(h)
+  }
+  const expectInvalid = (h: string) => {
+    expect(() => ensureValidAtUri(h)).toThrow()
+    expect(() => ensureValidAtUriRegex(h)).toThrow()
+  }
+
+  it('enforces spec basics', () => {
+    expectValid('at://did:plc:asdf123')
+    expectValid('at://user.bsky.social')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/record')
+
+    expectValid('at://did:plc:asdf123#/frag')
+    expectValid('at://user.bsky.social#/frag')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post#/frag')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/record#/frag')
+
+    expectInvalid('a://did:plc:asdf123')
+    expectInvalid('at//did:plc:asdf123')
+    expectInvalid('at:/a/did:plc:asdf123')
+    expectInvalid('at:/did:plc:asdf123')
+    expectInvalid('AT://did:plc:asdf123')
+    expectInvalid('http://did:plc:asdf123')
+    expectInvalid('://did:plc:asdf123')
+    expectInvalid('at:did:plc:asdf123')
+    expectInvalid('at:/did:plc:asdf123')
+    expectInvalid('at:///did:plc:asdf123')
+    expectInvalid('at://:/did:plc:asdf123')
+    expectInvalid('at:/ /did:plc:asdf123')
+    expectInvalid('at://did:plc:asdf123 ')
+    expectInvalid('at://did:plc:asdf123/ ')
+    expectInvalid(' at://did:plc:asdf123')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post ')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post# ')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post#/ ')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post#/frag ')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post#fr ag')
+    expectInvalid('//did:plc:asdf123')
+    expectInvalid('at://name')
+    expectInvalid('at://name.0')
+    expectInvalid('at://diD:plc:asdf123')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.p@st')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.p$st')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.p%st')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.p&st')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.p()t')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed_post')
+    expectInvalid('at://did:plc:asdf123/-com.atproto.feed.post')
+    expectInvalid('at://did:plc:asdf@123/com.atproto.feed.post')
+
+    expectInvalid('at://DID:plc:asdf123')
+    expectInvalid('at://user.bsky.123')
+    expectInvalid('at://bsky')
+    expectInvalid('at://did:plc:')
+    expectInvalid('at://did:plc:')
+    expectInvalid('at://frag')
+
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/' + 'o'.repeat(800))
+    expectInvalid(
+      'at://did:plc:asdf123/com.atproto.feed.post/' + 'o'.repeat(8200),
+    )
+  })
+
+  it('has specified behavior on edge cases', () => {
+    expectInvalid('at://user.bsky.social//')
+    expectInvalid('at://user.bsky.social//com.atproto.feed.post')
+    expectInvalid('at://user.bsky.social/com.atproto.feed.post//')
+    expectInvalid(
+      'at://did:plc:asdf123/com.atproto.feed.post/asdf123/more/more',
+    )
+    expectInvalid('at://did:plc:asdf123/short/stuff')
+    expectInvalid('at://did:plc:asdf123/12345')
+  })
+
+  it('enforces no trailing slashes', () => {
+    expectValid('at://did:plc:asdf123')
+    expectInvalid('at://did:plc:asdf123/')
+
+    expectValid('at://user.bsky.social')
+    expectInvalid('at://user.bsky.social/')
+
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post/')
+
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/record')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post/record/')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post/record/#/frag')
+  })
+
+  it('enforces strict paths', () => {
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/asdf123')
+    expectInvalid('at://did:plc:asdf123/com.atproto.feed.post/asdf123/asdf')
+  })
+
+  it('is very permissive about record keys', () => {
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/asdf123')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/a')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/%23')
+
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/$@!*)(:,;~.sdf123')
+    expectValid("at://did:plc:asdf123/com.atproto.feed.post/~'sdf123")
+
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/$')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/@')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/!')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/*')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/(')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/,')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/;')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/abc%30123')
+  })
+
+  it('is probably too permissive about URL encoding', () => {
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/%30')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/%3')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/%')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/%zz')
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post/%%%')
+  })
+
+  it('is very permissive about fragments', () => {
+    expectValid('at://did:plc:asdf123#/frac')
+
+    expectInvalid('at://did:plc:asdf123#')
+    expectInvalid('at://did:plc:asdf123##')
+    expectInvalid('#at://did:plc:asdf123')
+    expectInvalid('at://did:plc:asdf123#/asdf#/asdf')
+
+    expectValid('at://did:plc:asdf123#/com.atproto.feed.post')
+    expectValid('at://did:plc:asdf123#/com.atproto.feed.post/')
+    expectValid('at://did:plc:asdf123#/asdf/')
+
+    expectValid('at://did:plc:asdf123/com.atproto.feed.post#/$@!*():,;~.sdf123')
+    expectValid('at://did:plc:asdf123#/[asfd]')
+
+    expectValid('at://did:plc:asdf123#/$')
+    expectValid('at://did:plc:asdf123#/*')
+    expectValid('at://did:plc:asdf123#/;')
+    expectValid('at://did:plc:asdf123#/,')
   })
 })
