+++ packages/uri/tsconfig.json
@@ -6,4 +6,8 @@
     "emitDeclarationOnly": true
   },
   "include": ["./src","__tests__/**/**.ts"],
+  "references": [
+    { "path": "../identifier/tsconfig.build.json" },
+    { "path": "../nsid/tsconfig.build.json" },
+  ]
 }
